+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sse-auth/react",
-  "version": "1.2.0-alpha-1",
+  "version": "1.2.0-alpha-2",
   "main": "dist/index.cjs.js",
   "module": "dist/index.es.js",
   "types": "dist/index.d.ts",
