--- src/providers/Facebook/index.ts
@@ -1,3 +1,0 @@
-export * from "./types";
-export * from "./useFacebook";
-export * from "./FacebookLogin";
