--- src/providers/Github/index.ts
@@ -1,3 +1,0 @@
-export * from "./useGithub";
-export * from "./types";
-export * from "./GithubLogin";
