+++ src/providers/index.ts
@@ -3,4 +3,4 @@ export * from "./Github";
 export * from "./Facebook";
 export * from "./Auth0";
 export * from "./BattleDotNet";
-export * from "./types/types";
+export * from "./types";
