+++ src/providers/types.ts
@@ -1,0 +1,9 @@
+export interface UserProps {
+  [key: string]: any;
+}
+
+export interface ResponseProps {
+  error: Error | null | unknown;
+  accessToken: string | null | object;
+  userData: UserProps | null;
+}
