--- src/providers/types/types.ts
@@ -1,3 +1,0 @@
-export interface UserProps {
-  [key: string]: any;
-}
